refactor(cli): split remote ops into package modules
